pyproject.toml
[tool.poetry]
name = "toothless-bot"
version = "1.0"
description = ""
authors = ["Skye Im <skye.im@nyu.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
arpeggio = "^1.9.2"
tymora = {git = "git@github.com:neersighted/tymora.git"}
"discord.py" = "^1.6.0"
dateparser = "^1.0.0"

[tool.poetry.dev-dependencies]
mock = "^4.0.2"
pytest = "^5.4.3"
pytest-asyncio = "^0.14.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
